Separate AuditorAware configuration
