+++ src/main/java/org/isf/utils/db/JpaConfig.java
@@ -21,12 +21,9 @@
  */
 package org.isf.utils.db;
 
-import java.lang.reflect.InvocationTargetException;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
@@ -49,22 +46,4 @@ public class JpaConfig {
 		}
 		return () -> Optional.of("defaultAuditor");
 	}
-
-	@Bean
-	@ConditionalOnClass(name = "org.isf.security.ApiAuditorAwareImpl")
-	public AuditorAwareInterface auditorAwareCustomizer() {
-		try {
-			Class< ? > customAuditorAwareImplClass = Class.forName("org.isf.security.ApiAuditorAwareImpl");
-			return (AuditorAwareInterface) customAuditorAwareImplClass.getDeclaredConstructor().newInstance();
-		} catch (ClassNotFoundException | IllegalAccessException | InstantiationException | NoSuchMethodException | InvocationTargetException e) {
-			return new DefaultAuditorAwareImpl();
-		}
-	}
-
-	@Bean
-	@ConditionalOnMissingBean(AuditorAwareInterface.class)
-	public AuditorAwareInterface defaultAuditorAwareCustomizer() {
-		return new DefaultAuditorAwareImpl();
-	}
-
 }
